Move authentication mechanisms to using instantiated implementations
An AuthMechConfig now takes a fully instantiated implementation, instead
of the class + scheme. This allows an implementation to have
dependencies injected into it, e.g. an \OCP\IDB for database operations.
The downside is that authentication mechanisms now 'look' different to
backends, which is an inconsistency.

When a StorageConfig is being prepared for mounting, the authentication
mechanism implementation (IMechanism) has manipulateStorage() called,
which inserts the relevant authentication method options into the
storage ready for mounting.

An AuthMechConfig is still required to define the frontend
configuration, such as parameters, dependencies and custom JS files.